pyproject.toml
[tool.poetry]
name = "country-etl-project"
version = "0.1.0"
description = ""
authors = ["Guilherme <guiga.arruda@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
pandas = "^2.2.3"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
